+++ src/Repositories/UserRepositoryEloquent.php
@@ -4,6 +4,8 @@ use Prettus\Repository\Eloquent\BaseRepository;
 use Prettus\Repository\Criteria\RequestCriteria;
 use App\Entities\User;
 use Config;
+use Log;
+use Exception;
 
 /**
  * This file is part of Entrust GUI,
@@ -44,6 +46,9 @@ class UserRepositoryEloquent extends BaseRepository implements UserRepository
         $defaults = ['roles' => []];
         $attributes = array_merge($defaults, $attributes);
         $model = parent::create($attributes);
+        if( ! in_array('Esensi\Model\Contracts\HashingModelInterface', class_implements($model))) {
+            throw new Exception("User model must implement Esensi\Model\Contracts\HashingModelInterface. Revert to 0.3.* or see upgrade guide for details.");
+        }
         $model->roles()->sync($attributes['roles']);
         return $this->parserResult($model);
     }
@@ -60,7 +65,19 @@ class UserRepositoryEloquent extends BaseRepository implements UserRepository
     {
         $defaults = ['roles' => []];
         $attributes = array_merge($defaults, $attributes);
-        $model = parent::update($attributes, $id);
+        $model = $this->find($id);
+        if( ! in_array('Esensi\Model\Contracts\HashingModelInterface', class_implements($model))) {
+            throw new Exception("User model must implement Esensi\Model\Contracts\HashingModelInterface. Revert to 0.3.* or see upgrade guide for details.");
+        }
+        if( ! array_key_exists('password', $attributes)) {
+            $model->fill($attributes);
+            if(Config::get('entrust-gui.confirmable') === true) {
+                $model->password_confirmation = $model->password;
+            }
+            $model->saveWithoutHashing();
+        } else {
+            $model = parent::update($attributes, $id);
+        }
         $model->roles()->sync($attributes['roles']);
         return $this->parserResult($model);
     }