+++ views/permissions/index.blade.php
@@ -3,7 +3,9 @@
 @section('heading', 'Permissions')
 
 @section('content')
-<a class="btn btn-labeled btn-primary" href="{{ route('entrust-gui::permissions.create') }}"><span class="btn-label"><i class="fa fa-plus"></i></span>{{ trans('entrust-gui::button.create-permission') }}</a>
+<div class="models--actions">
+    <a class="btn btn-labeled btn-primary" href="{{ route('entrust-gui::permissions.create') }}"><span class="btn-label"><i class="fa fa-plus"></i></span>{{ trans('entrust-gui::button.create-permission') }}</a>
+  </div>
 <table class="table table-striped">
     <tr>
         <th>Name</th>